v8.2.2: Wait for mission patient controls
